+++ src/core/router/CHANGELOG.md
@@ -13,7 +13,7 @@ Changelog
 
 #### :bug: Bug Fix
 
-* [Get rid of redundant router transition when restoring the page from bfcache in safari in browser engine](https://github.com/V4Fire/Client/issues/552)
+* [Get rid of a redundant router transition when restoring the page from bfcache in safari in browser engine](https://github.com/V4Fire/Client/issues/552)
 
 ## v3.0.0-rc.182 (2021-04-28)